refactor(datastore): Move conflict resolution tests to their own file (#12728)
